requirements.txt
aniso8601==9.0.1
arrow==1.3.0
attrs==24.2.0
av==11.0.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blackboxai==2.6
blinker==1.8.2
briefcase==0.3.15
build==1.0.3
CacheControl==0.14.0
cattrs==24.1.2
certifi==2024.8.30
cffi @ file:///private/tmp/cffi-20231211-40239-j25k6o/cffi-1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
cleo==2.1.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cookiecutter==2.4.0
coremltools==8.1
crashtest==0.4.1
ctranslate2==4.5.0
cycler==0.12.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
dmgbuild==1.6.1
dnspython==2.6.1
ds-store==1.3.1
dulwich==0.21.7
email_validator==2.1.1
et-xmlfile==1.1.0
faster-whisper==1.0.1
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.4.0
git-filter-repo==2.47.0
gitdb==4.0.10
GitPython==3.1.37
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httptools==0.6.1
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.4
imageio==2.35.1
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.3
joblib==1.4.2
jupyter_core==5.7.2
kaldialign==0.9.1
keyring==24.3.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lxml==5.3.0
mac-alias==2.2.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.3.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.24.3
oauthlib==3.2.2
onnx==1.17.0
onnxruntime==1.16.0
openai-whisper @ git+https://github.com/openai/whisper.git@fc5ded7d9045c693692f13853857c3f8baea3a7b
opencv-python==4.10.0.84
openpyxl==3.1.2
opentelemetry-proto==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pexpect==4.9.0
pillow==10.3.0
pipenv==2024.0.3
pkginfo==1.10.0
platformdirs==4.2.1
pluggy==1.3.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pyaml==24.9.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic_core==2.23.4
Pygments==2.17.2
pygorithm==1.0.4
pylance==0.18.0
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.1.0
pytest==7.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.8.1
regex==2024.4.28
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.6.0
rsa==4.9
rubicon-objc==0.4.6
safetensors==0.4.5
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
std-nslog==1.0.3
sympy==1.12
tenacity==8.2.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metal==1.1.0
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
toga-cocoa==0.3.1
toga-core==0.3.1
tokenizers==0.15.2
toml==0.10.2
tomli_w==1.0.0
tomlkit==0.12.4
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.46.3
travertino==0.3.0
trove-classifiers==2024.4.10
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.9.0
urllib3==2.0.4
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.26.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.2
whisper-live==0.5.1
wrapt==1.16.0
xattr==1.1.0
XlsxWriter==3.2.0
yarl==1.9.4
yaspin==3.0.2
zipp==3.18.1
